requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
gitdb==4.0.5
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.4
pandas==1.1.5
Pillow==8.0.1
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-telegram-bot==13.1
pytz==2020.4
schedule==0.6.0
scipy==1.5.4
seaborn==0.11.1
six==1.15.0
smmap==3.0.4
tornado==6.1
tzlocal==2.1
Werkzeug==1.0.1
